backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
httpx
python-dotenv
aiofiles
